:recycle: refactor: Include response entity to serve more precises http status to the user, and replace model usage for dto

+++ src/main/java/org/otavio/cadastrodeninjas/Missoes/Controller/MissionsController.java
@@ -1,8 +1,13 @@
 package org.otavio.cadastrodeninjas.Missoes.Controller;
 
 
+import org.apache.coyote.Response;
+import org.otavio.cadastrodeninjas.Missoes.Dto.MissionsDTO;
 import org.otavio.cadastrodeninjas.Missoes.Models.MissionsModel;
 import org.otavio.cadastrodeninjas.Missoes.Service.MissionService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,29 +23,51 @@ public class MissionsController {
     }
 
     @PostMapping("/create")
-    public MissionsModel criarMissoes(@RequestBody MissionsModel misson) {
-        return missionService.criarMissao(misson);
+    public ResponseEntity<String> criarMissoes(@RequestBody MissionsDTO missonDTO) {
+        MissionsDTO missionDTO = missionService.criarMissao(missonDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Missão criada com sucesso " + missionDTO.getName() +
+                "ID: " + missionDTO.getId());
     }
 
     @GetMapping("/show")
-    public List<MissionsModel> listarTodasMissoes() {
-        return missionService.listarMissoes();
+    public ResponseEntity<List<MissionsDTO>> listByID() {
+        List<MissionsDTO> missions = missionService.listarMissoes();
+        return ResponseEntity.ok(missions);
     }
 
     @GetMapping("/show/{id}")
-    public MissionsModel listarMissoesPorId(@PathVariable Long id) {
-        return missionService.listarMissoesPorID(id);
+    public ResponseEntity<?> listMissionByID(@PathVariable Long id) {
+        MissionsDTO missionsDTO = missionService.listarMissoesPorID(id);
+
+        if (missionsDTO != null) {
+            return ResponseEntity.ok(missionsDTO);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("A missão com o ID: " +
+                id + " não foi encontrada nos registros, " +
+                "tente novamente.");
     }
 
     @PutMapping("/alterar/{id}")
-    public MissionsModel alteraMissao(@PathVariable Long id, @RequestBody MissionsModel mission) {
-        return missionService.updateMissions(id, mission);
+    public ResponseEntity<?> updateMissionByID(@PathVariable Long id, @RequestBody MissionsDTO missionDTO) {
+        MissionsDTO mission = missionService.updateMissions(id, missionDTO);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("A missão com o id: " +
+                id + " não foi encontrada nos registros, tente novamente ");
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteMission (@PathVariable Long id) {
-         missionService.deleteMissionsByID(id);
+    public ResponseEntity<String> deleteMissionByID(@PathVariable Long id) {
+        if (missionService.listarMissoesPorID(id) != null) {
+            missionService.deleteMissionsByID(id);
+            return ResponseEntity.ok("Missão deletada com sucesso " + "ID: " + id);
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não foi possivel deletar, ID: " + id +
+                ", não foi encontrado nos registros, tente novamente.");
+
     }
-
-
 }
